pyproject.toml
[project]
name = "invigo-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "filelock>=3.18.0",
    "jinja2>=3.1.6",
    "msgspec>=0.19.0",
    "natsort>=8.4.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.2",
    "schedule>=1.2.2",
    "tornado>=6.5.1",
]
